Refactor: Move client code to modular structure with separate style and HTML files
